Add support for multiple data source directories
In order to allow the program to run on multiple machines, it needs to look
for Garmin data in multiple directories. Directories that don't exist
should be skipped without action. This was implemented by expanding the
source arrays in bin/g2p-ini.sh, as explained in the source file comments.